requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
dateparser==1.2.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
pycryptodome==3.23.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2024.11.6
requests==2.32.4
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
